Update Distribution.Version
Change the Version type in Distribtion.Version to be a combination of
a branch ([Int]) and a list of arbitrary tag strings.  This is
intended to be a generic version type, where the interpretation of
ordering is application-dependent.

Distribution.Version also has lots of version-related stuff that is
specific to packages: version ranges (the VersionRange type), and
concrete show/parse functions for package versions.

20040427155356-3ed52-3850624803764bcd2419b1a3f68e259579e649b4.gz