Remove -DDEBUG from OPTIONS pragmas
If you don't actually have HUnit installed, it's a pain in the butt
to pass -i parameters or -package-conf parameters to ghci when you
just want to check something in a module or another. Makes more sense
to require -DDEBUG on the command line. Also, I changed the Makefile
to define -DDEBUG when building moduleTest, so that should still work.

20040725120303-153e8-a5077cbb0a49a24b3a514ec1bd2ef80720346205.gz