fixed checking for unknown args
Long story:

It seems that Marvin removed my beautiful and powerful "getOpt"
alterations to use the System.Console one, which I had already deemed
to be non-beautiful because of the errors it spews when faced with
arguments it doesn't understand.

I think Marvin misunderstood the API docs that seem to suggest that
the "commands" that get passed back from getOpt are the non-matched
arguments.  In fact, non-matched arguments get returned as errors
(ewwwww).

So I had to re-apply my GetOpt module, and alter the parser
accordingly.

Also, since it's not OK to pass random arguments all the time (just
when you have user hooks), I cleaned up calls to no_extra_flags to
only get called when there's nothing sensible to do with unknown
arguments.


20050106071635-6d90f-3f6514c1b7c55a4720b1180856d6147aef2d8d24.gz