look in libraries source directory for executable modules
When building executables, one typically wants to build them based on
the library that's contained in the cabal package.  This was difficult
if both the executables and the library used different
hs-source-dir's.  This patch causes executables to look at the
library's hs-source-dir for modules when preprocessing and building.

This is instead of building and locally installing a package and using
the -package flag to build the executable, since that's more
complicated, and won't necessarily work if the library isn't designed
to be installed in-place.


20050108215448-6d90f-120c3bcb8d9e9084342c2e6a543b016b5b49c8a0.gz