added ./setup haddock cmd. works on exposed-modules
It requires both haddock and hscpp; Fix to not require cpp if it's not
needed.

It runs the cpp over the files first and unlits .lhs files.

Along the way, cleaned up the command-line parser to re-use code for
commands with no arguments.

Changed the type of withLib to be more general


20050110040341-6d90f-3b21398cd9e9d41d6e0a05f1a92a1d91bf6434df.gz