Krasimir implements Support for -v/--verbose option
The option supports levels from 0 to 5 but currently the library
distinguishes only zero and nonzero levels. The option is supported
only for build and haddock commands for now. I plan to add support for
level 5 of verbose. In this level all external commands will be
executed with -v option too.


20050116194050-6d90f-d967bcf9b99329eb91765b0484ecfc3f3924f6c2.gz