added programatica support (may remove)
On a whim (and since I randomly ran into Thomas Hallgren on the train
today) I added support for Programatica, via pfesetup.  It doesn't
probably do quite what one would want it to do since it preprocesses
everything and slaps it into dist/tmp, but it may be interesting
nonetheless, and we can clean it up later if anyone thinks we should
keep it around.

It was very easy to add and looks a lot like the haddock command.  Try
it with ./setup pfe.  It doesn't do anything too interesting to the
cabal sources yet since it can't find a bunch of modules.  It would be
great if someone added all the fptools modules to it :)


20050204033034-6d90f-4cfcdd88a85943f8b381cf9c6023fb92ecb470f5.gz