implemented ranlib execution during installation
This is necessary in Mac OS X and on Solaris.

It is possible that this might cause problems for systems that use
setup copy to build and then use cp to install files, since the
library modification time will change.  I think this is only the case
on Mac OS X and Solaris.


20050502050154-6d90f-e02edff32800b4053d3e2c3c309fdba44c8bbaca.gz