fixed bug: executable build location had changed
Since we don't have hs-source-dir anymore, the location for building
executables had changed.  However, install was still looking in the
old location.  Fixed this.

Fixed another bug where hs-source-dir for executables was [".","."]
where this duplication caused a problem in locating the Main module.
Not sure quite where this came from, but now for combining fields, we
always nub, so this should be a robust fix.

Fixed build failures in testing code.

20050515074133-6d90f-7d51fe4dcac1d7f3b9edee94243686e112953637.gz