modified install --user behavior to include configure --user
If the user configures with --user, that means that they are willing
to satisfy dependencies from the user databaes.  If that's the case,
then it really only makes sense to install the package as --user as
well, since if user A installs configures / builds the package with
--user, then installs it as root, user B may attempt to use the
package and not be able to fulfill the dependencies.  Therefore, if
the package is configured as --user, then we also will call ghc-pkg
with --user.

Of course, in this case:

./setup configure --user
./setup install --global

Cabal will accept the user's instructions to install this globally,
however ill conceived.  This changes the type of the install hook,
btw, since we now need to be able to tell the difference between
passing the --global flag and not passing any flag.

Any objections to this change?

20051115060943-6d90f-86e0eb0c2079fa6389e45c365b334b2c010b4ced.gz