Add xargs function and use it when linking.
When using GHC's --split-objs we end up with lots of files to link.
This can mean overflowing the maximum length of the command line
when invoking ar or ld. On windows the maximum length is 32k. On
other systems it's not a great deal more. GHC currently deals with
this problem by using xargs. This patch does more or less the same.

20060818010449-9561d-c7637504e20715d411bc9dd161b4f15b7359c577.gz