Change default installation dirs for executables on Windows:
before:

  bindir  = $prefix\$pkgid
  datadir = $prefix

now:

  bindir  = $prefix\Haskell\bin
  datadir = $prefix\Haskell

datasubdir is still $pkgid, so data files for an exectuable will be
installed in $prefix\Haskell\$pkgid.

the idea is that all executables are installed in the same directory
by default, so setting PATH is easier.  Also, data files belonging to
Haskell programs are grouped together under $prefix\Haskell.

20060822134232-760e2-6c3375a2baeb50f303ac36db34fb5dbf06a2529c.gz