Be cleverer about guessing hc-pkg name and location.
So it now works if you say:
./setup configure --with-compiler=ghc-6.5
ie specifying a path-relative name rather than an absolute path.
We then look for hc-pkg in the same dir as where we found the compiler.
If the compiler appears to have a version suffix then we additionally
and preferentially look for hc-pkg with that same version suffix.
(I'm not sure that bit works if you've got a .exe suffix, perhaps a
windows person could try it / take a look)

20060823105412-9561d-1879e0da05b5a2edd69c4a2d1d838fe74d5e7cdb.gz