Refactorings only
Here are a batch of refactorings to clean up parsing and parts of the
simple build system.  This patch originated in a patch sent to
cabal-devel@haskell.org with an intial implementation of
configurations.  Since then we decided to go a different route with
configurations, so I have separated the refactoring from the
configurations patch.

At this point, 2 tests fail for me, but I get the same 2 failures
without this patch.

20070114203741-760e2-c22c91d4f0406c5b9b8a6e0e7882651f2172cc32.gz