Put pre-processed source into the dist/build dir rather than src dirs
This is generally just a nicer thing to do, we should probably aim to
not write any files into the source tree at all.
The main change is in the preprocessModule function. It now takes an extra
arg which is the destination directory. For now I'm passing the buildDir,
but we could consider putting pre-processed files into a separate fir
from where the .o and .hi files end up.
To work out the correct destination file we need to know not only the source
file but which of the search dirs it was found in, since the relative file
name will be the name of the source file relative to the search dir it was
found in, not the name relative to the top of the source tree. This is so that
we will be able to find the pre-processed .hs file just by adding dist/build
to the sources search path when we compile (eg with -i for ghc).
This almost certainly breaks the sdist thing where pre-processed files get
included into the tarball. So that'll need looking at.

20070422095631-9561d-cd4494956cf925374d9d82b62e617a7f895b3fbb.gz