call c2hs using the more detailed info
we use --output-dir=dist/base
and --output=<file relative to the search dir it was found in>.hs
This actually depends on a patch in c2hs to make it treat --output-dir
in the way we want. That patch will be forthcomming soonish.
But the point is:
c2hs --output-dir=dist/base --output=Foo/Bar.hs src/Foo/Bar.chs
will generate dist/base/Foo/Bar.hs and also dist/base/Foo/Bar.h
but inside the .hs file it'll reference Foo/Bar.h so when we compile the
.hs file we have to -Idist/base
Clear as mud?

20070422102617-9561d-6a0f663bb86dc977c1fceb31c9a6602ed36a7996.gz