Make a proper verbosity type, rather than using Int values
Hopefully this will make it easier to get better verbosity consistency.

We could, by changing only Distribution.Verbosity, use
"type Verbosity = Int" for now to give users of the library a chance to
catch up, but the upcoming Cabal release seems like a good opportunity
to cram in as much of the interface-changing stuff that we want to do
as we can. I think the added benefit of a slow switch would be very low
indeed.

20070514165514-3fd76-86a418e5a48690c253197dbb76ae5d10d1ffd2ea.gz