Add version field to Program and add findProgram(AndVersion) utilities
findProgram and findProgramAndVersion construct Program values
The latter makes it easier to parse output of --version calls.
The idea is that we should carry the version number along with the program
if we know it, so we can later decide version-dependent args without having
to do more IO to find the program version, eg see use_optP_P & haddockVersion.

20070802172941-adfee-db56603a9b3e1109e8d6a88fca0815b56b1a5f71.gz