Add compilerExtensions field to Compiler and make each compiler fill it in
It's just a list of supported extensions and the corresponding compiler flags.
For most compilers this is currently just a static list. For ghc 6.7 and above
we query ghc to find the list of language extensions it supports.
In each case the code has moved out into the compiler-specific modules and the
core code treats it generically.
The extensionsToFlags function has been split into two:
extensionsToFlags which now returns the flags for the supported extensions and
unsupportedExtensions which does what it says it does. This is because the two
roles of the previous function were always used separately, never together.

20070807170653-adfee-950ae74a3dcee5f3a92309dfa76560ff48a9d650.gz