Rewrite the Program abstraction and the ProgramConfiguration database
Also make the follow on changes to everything that uses Program.
The notion of a program is now split into the abstract notion of a program
that we know about and might be able to configure, and a configured program
that we can actually run. The ProgramConfiguration database is similarly
split. We still keep user-supplied loation and arguments and use them when
we configure programs. The abstract Program now has functions to search for
the program on the system and for finding the version number. This allows
for more generic configuration of programs.

20070817033841-adfee-78d02fa0ab91cfde75e2b2adaebba0f79ccf6d49.gz