Make checking program versions not produce error spew
Redirect stderr to /dev/null (or NUL on windows) and throw an exception
like rawSystemExit does rather than calling die which prints an unhelpful
message before throwing an exception. It's not an overall failure to not
be able to establish the version of a program, since we check later if we
actually have any requirement for any particular version anyway.
So for example this makes running configure with an old version of c2hs
that doesn't grok --numeric-version not produce unhelpful output that
make people think that configure failed.

20070830144312-adfee-c6f40a8553d0dc2419a808d83f60cb510056d7bb.gz