Generalise build --ghc-option flag to --prog-option for any prog
and remove makefile --ghc-option flag as it's not used (as far as I can see).
It works generically by updating the ProgramConfiguration rather than the
previous ad-hoc implementation. This feature is generally only useful for
hackers who want to pass extra args to a progam during the build step
(probably as a one-off) without having to reconfigure.

20070903110658-adfee-cbec1b98b3842e680279f757add4c0fbca66434b.gz