Move expensive conviguration actions inside branch where they're used
We only need to configure the compiler and resolve any "Cabal-Version:"
dependency when we're actually going to compile the Setup.hs program.
We can otherwise save time by not calling ghc and ghc-pkg unnecessarily.
(I only noticed this because I've got >150 registered packages which
causes ghc and ghc-pkg to be very slow to start.)

20070917085834-adfee-41693d7f8f4e5152520817d144c607eb0cf029f3.gz