Change the handling of cpp & ghc flags for hsc2hs
The hsc2hs that comes with ghc uses ghc as the C compiler. This means we must
escape the real cc flags. It also means we can ask ghc to add extra include
dirs that we might need to find includes of dependent packages. This is a bit
of a hack. In the longer term it'd be better for Cabal to collect the include
dirs and cc options of dependent packages and to pass them explicitly itself.

20071016231652-adfee-e8bdf97173786057746ee09a0a8acc6536ad6f8d.gz