Figure out if hsc2hs is using gcc or ghc as it's C compiler
and pass the appropriate flags on the basis of knowing that.
This is a hack.
What we should do longer term is make hsc2hs always use gcc as it's C compiler
and have Cabal figure out the right flags to pass it, rather than using ghc
to pass on the appropriate flags to gcc.

20071017143108-adfee-41380bb344d3f2e4854e74b2b8a130ab20d00b76.gz