Replace lots of configBlahDir fields with just one InstallDirs
So instead of 10 fields like this:
configPrefix   :: Maybe FilePath,
...
we have just one compound one:
configInstallDirs   :: InstallDirs (Maybe FilePath)
Now that LocalBuildInfo and cabal-install's config use InstallDirs it makes
converting a lot easier since we can use fmap and combineInstallDirs.
This should also be easier to maintain as it has significantly reduced the
number of places in the code you have to change when you add/change a dir.

20071208182620-adfee-9cd852bf8795c7c00f3ac33733ca727786e16a5b.gz