Make all command line flags and flag collection types monoids
And specify command flags in a way that allows converting from or two strings.
So not only can we parse command lines into the various flag types, we can
convert back. This is crucial in cabal-install so that we can call Setup.hs
scripts with all the various args, and override certain flags in a typed way
rather than just representing them all as strings.

20071215193650-adfee-bbd249904ed63af01a64a84fd19b11d4d271ce48.gz