Add correct^H^H^H^H British English spelling of optimisation
For the --enable/disable-optimi(z|s)ation flag, allow the British English
spelling of the flags as an alias. The current one is still the one
displayed, and it does not list bith in the --help text, though the command
completion would list both.
Perhaps the default should be on the basis of the current locale :-).

20071217133533-adfee-1dc9e6af6b7b0cf6297c3204a4c6467095cf7376.gz