Make --user --global change the default prefix
Previously --user and --global didn't mean what it looks like they mean,
instead of doing a per-user or a global installation they just meant to
register in the per-user or global package databases (and to allow
dependencies to be satisfied from the per-user or global db)
With this patch --user / --global means to do a per-user or global install
so in addition to the package db differences it also sets the default 
installation prefix. This prefix can of course still be overridden using the
--prefix= flag. The global prefix is as before, the default per-user prefix
is $HOME/.cabal on Unix and on Windows it's something like:
C:/Documents And Settings/$user/Application Data/cabal
(using getAppUserDataDirectory "cabal"). This is the per-user prefix that
cabal-install currently uses. We can change it if people think it's not good.

20080110193830-adfee-cca3a60822d35b378601be179b978bb5047521b9.gz