Move the license check from a sanity check to a QA check
Also extend the check. It's now a QA error to use a license that is empty
or AllRightsReserved. It's a warning to not specify a license file. It's an
error for the license file to be given but not exist.
Note that QA errors are not fatal for generating tarballs. But they will become
fatal errors for uploading to the public hackage server. It will still be
possible to use AllRightsReserved or another license on a private hackagedb.

20080121001722-adfee-68150abaf3718d5aaec77a08f4a64bf32677d33c.gz