Simplify smartCopySources using findFileWithExtension' and copyFiles
Previously it used moduleToFilePath2 which returns all the locations in which
a source file is found, in different search dirs and with different extensions
and it just collected all these locations. This is totally bogus since we do
not want to copy multiple source files per-module, only one of them could have
been used to be pre-processed into the final source module. This is one
example of how almost every use of moduleToFilePath is convoluted or broken.
This function it still used for too many different purposes, witness the two
flags to customise its behaviour which at all call sites are constants.

20080130000129-adfee-1fae8990da880bfa5678eab76362794488879a9b.gz