Stop using smartCopySources in SrcDist
This was the old module using it in with preserveDirs = True.
Anyway, we should be using smartCopySources for each lib and exe, we should be
collecting all the files to copy together into one list, then eliminating
duplicates and doing a single copyFiles operation. Yes, there can be
duplicates because for example exes can specify the libbrary's modules or 
even the other exe's main module as other-modules.

20080131114143-adfee-e308fbf45fc0dc8b8e4ec177e44d00bae18b22be.gz