Adjust unlit's behaviour on continuing comments
This behaviour is not set in stone, there is not full agreement yet.
The previous code was deciding wether to continue a comment based on wether
the intervenienting lines were completely blank or had at least one space in.
The other suggestion was to use a '.' to continue a comment line.
Instead, this change uses a single blank line between comments to continue
a comment and anything else terminates it with a blank line.

20080131142901-adfee-84d75c40438951454139cc196d661ddd2c76d953.gz