Fudge comment indentation in unliting to work with haddock
The rule is, if we see any bird track style code then we will indent all
comments by two spaces so that it should line up with the code. Otherwise
we use no indentation so that it'll work with latex style literate files.
This makes it work for takusen (once you change the '.' lines to be blank).

20080203042308-adfee-10f1913e7470eb65c778e9a3d0f7741b3dc2c435.gz