Replace moduleToFilePath with findFileWithExtension
moduleToFilePath encouraged bad error handling and becuase it was specific
to module names it made it hard to treat modules and main.hs files uniformly.
This is one step towards using the same code paths for modules and main.hs
files so that main.hs files always work properly with pre-processors etc.

20080206014554-adfee-03782fce74dc09712e715b4bd1b08a8b0cd37139.gz