Fix a bug where the destdir prefix got included in the prog prefix and suffix
Previously all members of InstallDirs were paths so it was safe to prefix
them all with the destdir, but now we've also got these two non-paths, so
we have to not apply the prefix to them. Perhaps the program prefix and suffix
should just be PathTemplates on their own and not part of InstallDirs.
Reported by MarcWeber while hacking on Haskell packages for Nix.

20080208192613-adfee-01a8d6a9bd175afd038264bd2943e2689acf496f.gz