Unify sanity and other checks and separate IO checks
And rename from QA so we no longer pretend to assure anything :-)
We now distinguish different kinds of tests with a PackageCheck type so the
caller gets to decide which kinds of checks are relevant to the context.
This PackageCheck type should be extensible if we want to add more kinds or
more info besides a brief human readable explanation.

20080211003830-adfee-57699c5ec277e6a5b2be3890dc4624bbaf3663f0.gz