Detect broken and inconsistent package deps
We now check for packages that are broken due to their dependencies having
been unregistered. We fail and print a fairly sensible message in this case.
We also check for inconsistent dependencies and give a warning saying which
packages are depending on inconsistent versions of a third. This is a warning
not an error because it does not always lead to failure. Hopefully it'll help
people who are otherwise just running into random compile errors.
This fixes ticket #220.

20080219000139-adfee-d134fd30bd059c384c5c3b980322771265eb19ce.gz