Use the InstalledPackageIndex rather than calling ghc-pkg in Haddock module
Previously we made two calls to ghc-pkg per-dependent package which starts
to get very slow once you have 100's of installed packages and the package
you're building depends on several of them. Now we use the cached information
about the installed packages. Also refactored the code to generate the haddock
package flags so we can export it and use it elsewhere more easily.

20080219125050-adfee-6308f007c172f8ca0cda3c9ae0e592bf515abb5a.gz