Check for invalid UTF8 when parsing .cabal files
This assumes a permissive UTF8 decoder has inserted '\xfffd' as a replacement
character. We should check strictly for errors when we decode instead, though
it's nice to do that where we have some kind of error reporting infrastructure
which is why I've added it to the parser for the moment. The current error
message is not too bad, as it reports the line number.

20080225132824-adfee-8765660b1b053a2e5769b1ab00eb41c5b690ed45.gz