Reformat License declaration and add related TODOs
The real difficulty with changing the license stuff is the parser in current
and older cabal versions is very fragile, it parses using Read so only
recognises those exact enum values and if parsing fails, then parsing the
whole .cabal file fails. We really need to make this more permissive for all
the enumerations so that we can make changes in future.

20080229224932-adfee-1373bcf29230d68ba5895f2f65a98feecc386db5.gz