Reorganise and prune tests/
Split them into unit tests and system tests. Also remove the rather large
bundled HSQL and HUnit packages. We can just depend on HUnit for the tests
rather than bundling it. HSQL is very large for a system test and we can
just as easily get it and others from hackage.

20080310183938-adfee-483577c4950840a1cf489f9261d3d10f7d53fc50.gz