Use the same ReadP for all compilers, remove CPP hacks
If we're bundling a whole copy of ReadP then why bother trying to use
the version from the base package, especially when that requires hacks
to use the H98 version with some compilers and the non-H98 version in
base. It just makes testing harder because we're using different versions
with different compilers.
As a bonus, hugs should no longer need the -98 flag to use Cabal.
Added all the type signatures back in and fixed some warnings.

20080310191422-adfee-0d58dbddf7be480cbe2360613753542d76a69714.gz