Change mkLibName and friends to take a package id and not take the dir
So now it really gives just the lib name and not its location as well.
It turned out that we more often want just the name since we're copying
the same named file from one directory to another.
Taking a PackageIdentifier means we do the conversion to string in one
place rather than all over the place.
It also means we're chaning the lib names for nhc98 to include the lib
version number. This should not make any real difference to nhc98.

20080317221259-adfee-704b064ac399c38c2fd0799e56883d73b786908f.gz