Refactor sdist code a bit
Separate the phases of sdist better so that they can be reused in
cabal-install. This part is similar to Andrea Vezzosi's patch.
Separate the snapshot code rather than mixing it all in with prepareTree.
Use a proper withTempDirectory which fixes one FIXME.
Better separate UI code from actions, move the notices to sdist.

20080318235916-adfee-bad5a9171bd437e62299a461559b1d290efdf9da.gz