Add Text class for displaying and parsing values
This is intended to replace all these showFoo read/parseFoo functions
we have all over the place. It's like the Read/Show classes but uses
a half decent parser and pretty printer:
class Text a where
  disp  :: a -> Doc
  parse :: ReadP a
This patch just adds the class and an instance for Bool, replacing
the parseBool function used in a few places. More to follow...

20080321015414-adfee-8cbac296771a4d782aea2afe218dd4b72f421b41.gz