Arch and OS names were previously allowed to contain "_-", restore that.
That is the arch and os strings in conditionals in .cabal files, like:
  if arch(x86_64)
Previously the parser used isAlphaNum c || (c `elem` "_-"). This is
crucial for arch names like "x86_64". So make the new parser for OS
and Arch types use the same string parser as before (there is a string
parse part and a separate classification of known OS/Arch values).

20080327004631-adfee-c36a10693a672ccdd950fd19c959d9f788d05afd.gz