Separate the OS/Arch classifiation used for different purposes
We have to classify System.Info.{os,arch} strings into the OS and Arch
enums. For that purpose we have to be quite permissive since there are
lots of ailases that the various Haskell implementations use. However
for parsing os and arch names in .cabal files we'd prefer to use
canonical names, though we do have to allow the couple aliases already
in common use.

20080327005321-adfee-9da906d611dce882b49cd5be58229c2233d50d0a.gz